Revert "альтернатива"
This reverts commit f6b445b6e19eade3a5a153a9b1c3ed130df78f0a.

+++ src/main/java/ru/gb/eshop/gb_eshop/controllers/AdminController.java
@@ -145,6 +145,9 @@ public class AdminController {
 
     @GetMapping("/product/delete/{id}")
     public String deleteProduct(@PathVariable("id") int id) {
+        if(orderRepository.findByProduct_id(id)!=null){
+            return "redirect:/admin";
+        }
         productService.deleteProduct(id);
         return "redirect:/admin";
     }
